+++ client/src/!custom/views/socket/ConnectionManager.vue
