+++ client/src/!custom/views/socket/ConnectionState.vue
